Move SyncDatabase to sub-package
