+++ server/src/main/sqldelight/sync/eu/pretix/pretixscan/scanproxy/sqldelight/sync/ProxyQueuedCheckIn.sq
